Bug fixation of observe_redds
The observe_redds C++ function was calling a at(i).year() in the Collection.cpp. This was fine as long as there was enough individuals in the collection. But eventually, if the number of salmon < number of redds, the at(i) mistake made crash the model... was long to understand ;-)